% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{caselaw}
\alias{caselaw}
\alias{cl_client}
\title{Class representing the case law API}
\description{
Class representing the case law API

Class representing the case law API
}
\section{Usage}{

\preformatted{
cl <- cl_client$new()
}
}

\section{Details}{


This class allows a user to interact with a case law API and provides
utility functions for GETs including cursor pagination and
the specifics of the case law query parameter structure

The constructor respects CASE_LAW_API_KEY environment variable if set, and will
use it to authorize all requests
}

\concept{R6 classes}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{cl_client$new()}}
\item \href{#method-raise_error}{\code{cl_client$raise_error()}}
\item \href{#method-add_auth}{\code{cl_client$add_auth()}}
\item \href{#method-add_url_params}{\code{cl_client$add_url_params()}}
\item \href{#method-GET_PAGES}{\code{cl_client$GET_PAGES()}}
\item \href{#method-GET}{\code{cl_client$GET()}}
\item \href{#method-GET_URL}{\code{cl_client$GET_URL()}}
\item \href{#method-GET_RESULT_URL}{\code{cl_client$GET_RESULT_URL()}}
\item \href{#method-clone}{\code{cl_client$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cl_client$new(base = "https://api.case.law", version = "v1")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-raise_error"></a>}}
\subsection{Method \code{raise_error()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cl_client$raise_error(res)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_auth"></a>}}
\subsection{Method \code{add_auth()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cl_client$add_auth()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_url_params"></a>}}
\subsection{Method \code{add_url_params()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cl_client$add_url_params(path, params)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GET_PAGES"></a>}}
\subsection{Method \code{GET_PAGES()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cl_client$GET_PAGES(
  path,
  writer = httr::write_memory(),
  parser = "parsed",
  limit = Inf
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GET"></a>}}
\subsection{Method \code{GET()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cl_client$GET(path, writer = httr::write_memory(), parser = "parsed")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GET_URL"></a>}}
\subsection{Method \code{GET_URL()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cl_client$GET_URL(url, writer = httr::write_memory(), parser = "parsed")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GET_RESULT_URL"></a>}}
\subsection{Method \code{GET_RESULT_URL()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cl_client$GET_RESULT_URL(url, writer = httr::write_memory())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cl_client$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
