% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{cl_get_case_opinions}
\alias{cl_get_case_opinions}
\title{Get case opinions}
\usage{
cl_get_case_opinions(case_id)
}
\arguments{
\item{case_id}{Case ID to retrieve}
}
\value{
Information on the case, 1 row per opinion
}
\description{
Get case opinions
}
