% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{cl_get_cases}
\alias{cl_get_cases}
\title{Get Cases}
\usage{
cl_get_cases(
  search = NULL,
  jurisdiction = NULL,
  court = NULL,
  reporter = NULL,
  docket_number = NULL,
  decision_date_max = NULL,
  decision_date_min = NULL,
  name_abbreviation = NULL,
  limit = Inf
)
}
\arguments{
\item{search}{A search term or phrase, case metadata and text is searched}

\item{jurisdiction}{A jurisdiction slug, eg. "ill" or "us". TODO: cl_get_jurisdictions()}

\item{court}{A court slug, see: \link{cl_get_courts}}

\item{reporter}{A reporter id. TODO: cl_get_reporters()}

\item{docket_number}{A string with the docket number}

\item{decision_date_max}{Max search date in YYYY-MM-DD}

\item{decision_date_min}{Min search date in YYYY-MM-DD}

\item{name_abbreviation}{Case name abbreviation as a string}

\item{limit}{Maximum number of search pages to return}
}
\value{
tidy metadata about matching cases, with an emphasis on context not publication info
}
\description{
Get Cases
}
